requirements.txt
Django==4.0  # django-celery-beat: Django <= 4.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-cleanup==6.0.0
django-celery-beat==2.3.0
django-pickling==1.0
django-redis==5.2.0

djangorestframework==3.13.1
drf-spectacular==0.23.1

psycopg2-binary==2.9.3
redis==4.3.4
Pillow==9.2.0
gunicorn==20.1.0
whitenoise==6.2.0
celery==5.2.7
gevent==21.12.0
requests==2.28.1
selenium==3.141.0
beautifulsoup4==4.11.1
fake_useragent==0.1.11
